1 file of code
"Stone Paper Scissor" is a simple Python program that allows users to play the classic game of "Stone Paper Scissors" against the computer. The game follows the traditional rules, where players choose to throw either "stone," "paper," or "scissors," and the winner is determined based on the following rules:

"Stone" beats "scissors"
"Paper" beats "stone"
"Scissors" beats "paper"
In this project, we have implemented a user interface that allows players to input their choice and see the result of the game. The computer's choice is determined randomly, using the built-in random module.

Overall, "Stone Paper Scissor" is a fun and easy-to-understand project that is perfect for those new to programming or Python. It provides a great introduction to basic concepts such as user input, random number generation, and conditional statements.